perf(mem): apply stringent MLX memory optimizations to avoid 8GB OOM explosion
- Switched PyTorch dataset loading back to np.memmap to prevent loading 1.7GB directly into RAM
- Re-architected microbatch compilation strategy: passed random matrices from host Python code into the compiled block instead of computing them inside mx.compile, resolving the graph trace evaluation explosion
- Bypassed astype(mx.float32) cast on massive 536MB logits array before cross-entropy, directly dropping the memory payload of calculating CE error
- Replaced is_special = (input_ids == 0) | (input_ids == 1) | ... with is_special = input_ids < 4 inside mask masking
- Configured trainer step evictions to run mx.clear_cache() synchronously every *single* step rather than every 100 steps to combat MLX Metal pool fragmentation

+++ undiff/diffusion/loss.py
@@ -13,7 +13,7 @@ def undlm_loss(model, noisy_ids, clean_targets, t_values, vocab_size, special_to
     B, T, V = logits.shape
 
     # Upcast logits to float32 for numerically stable log-softmax in cross entropy
-    logits_flat = logits.astype(mx.float32).reshape(-1, V) # (B*T, V)
+    logits_flat = logits.reshape(-1, V) # (B*T, V)
     targets_flat = clean_targets.reshape(-1) # (B*T,)
 
     # CE on every position in float32
